requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
pandas==2.3.3
openpyxl==3.1.2
reportlab==4.0.4
Werkzeug==2.3.7
scikit-learn==1.7.2
scipy==1.16.2
statsmodels==0.14.5
numpy==2.2.6
python-dateutil==2.9.0
pytz==2025.2